:FOR THE MODIFICATIONS AND NEW-EDITIONS
~1 FOR THE "clone"-FIELD OF THE "SmartParserState" IS WITH THE EDITS FOR THE CLONING OF THE
"path"-FIELD.

~2 FOR THE CONSTRUCTOR OF THE "SmartParserRule" IS WITH THE EDITS FOR THE NEW-RULE-TYPE:
STACK: "Regular Expression" AND "Regular Expression" OR "Function". WITH THIS EDITION OF THE
CODE IS WITH THE CERTIFICATION OF THE FIRST-RULE-MATCH WITH A "Regular Expression" OR
"Function". :TOKENIZATION+VERIFICATION=CERTIFICATION.

~3 FOR THE "match"-METHOD OF THE "SmartParserRule"S IS WITH THE EDITS FOR THE
NEW-RULE-TYPE-OPERATIONS.

~4 FOR THE CONSTRUCTOR OF THE "SmartParserSequence" IS WITH THE EDITS FOR THE CHECKING FOR
THE VOID-RULES WITHIN THE RULES-LIST.

~5 FOR THE STATE-TRACKING WITHIN THE "SmartParserSequence.match"-METHOD IS WITH THE EDITS
FOR THE SKIPPING OF THE TRACKING FOR THE SEQUENCES WITH THE LACK OF A NAME.

~6 FOR THE OR-OPERATION OF THE "SmartParserSequence.match"-METHOD IS WITH THE EDITS OF A
CORRECTION-CLAIM FOR THE NESTING OF THE "SmartParserSequence". FOR THE OLD-OPERATION OF THE
SEQUENCING IS WITH THE CANCELLATION OF THE PARENT-SEQUENCING FOR ANY FAILURE WITH THE
MATCHING OF A CHILD-SEQUENCE. FOR THE NEW-OPERATION OF THE SEQUENCING IS WITH THE
CONTINUATION OF THE PARENT-SEQUENCING THROUGH THE FINAL-SEQUENCE OF THE OR-OPERATION. FOR
THIS EDITION OF THE CODE IS FOR THE CORRECT-CLAIM OF A SEQUENCE-PARSER.